CRUDING comments

+++ backend/controllers/posts.js
@@ -6,27 +6,7 @@ const config = require('../config/config')
 const mongoose = require('../models/Post')
 const Post = mongoose.model('Post')
 const User = mongoose.model('User')
-
-function verifyToken(req, res, next) {
-    console.log("in verify...");
-    // Get auth header value
-    // when we send our token, we want to send it in our header
-    const bearerHeader = req.headers['authorization'];
-    console.log(bearerHeader)
-    // Check if bearer is undefined
-    if(typeof bearerHeader !== 'undefined'){
-        const bearer = bearerHeader.split(' ');
-        // Get token from array
-        const bearerToken = bearer[1];
-        // Set the token
-        req.token = bearerToken;
-        // Next middleware
-        next();
-    } else {
-        // Forbidden
-        res.sendStatus(403);
-    }
-}
+const verifyToken = require('./middleware')
 
 // /posts/  => Posts Index
 router.get('/',verifyToken, (req, res) => {
@@ -65,13 +45,20 @@ router.post('/new',verifyToken, (req, res) => {
 
 // posts/:id => Post delete
 router.delete('/:id',verifyToken,(req,res) => {
+    let userId = jwt.decode(req.token, config.jwtSecret)
     let postId = req.params.id
-    Post.deleteOne({ _id: postId }, (err, deletedPost) => {
-        if(err){ 
-            return res.status(400).json({err: "error has occured"})
+    Post.find({_id:postId}, (err,post) => {
+        if(userId.id == post[0].user){
+            Post.deleteOne({ _id: postId }, (err, deletedPost) => {
+                if(err){ 
+                    return res.status(400).json({err: "error has occured"})
+                }
+                res.json({data:deletedPost});
+            });
+        } else {
+            res.json(err)
         }
-        res.json({data:deletedPost});
-    });
+    })
 })
 
 module.exports = router
